+++ material/src/features/popover/misc/types.ts
@@ -1,0 +1,9 @@
+/**
+ * Side of trigger element at which popover panel is displayed
+ */
+export type PopoverSide = 'top'|'bottom'|'left'|'right';
+
+/**
+ * Alignment of popover panel on cross axis of side at which it is displayed
+ */
+export type PopoverAlignment = 'start'|'center'|'end';